text-to-tex-api/requirements.txt
flask
openai
google-cloud-secret-manager